+++ mod-source-record-manager-server/src/main/java/org/folio/dao/JobExecutionDaoImpl.java
@@ -7,12 +7,15 @@ import io.vertx.sqlclient.Row;
 import io.vertx.sqlclient.RowSet;
 import io.vertx.sqlclient.Tuple;
 import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.lang.text.StrSubstitutor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.dao.util.JobExecutionMutator;
 import org.folio.dao.util.PostgresClientFactory;
 import org.folio.dao.util.SortField;
+import org.folio.rest.jaxrs.model.DeleteJobExecutionsResp;
 import org.folio.rest.jaxrs.model.JobExecution;
+import org.folio.rest.jaxrs.model.JobExecutionDetail;
 import org.folio.rest.jaxrs.model.JobExecutionDto;
 import org.folio.rest.jaxrs.model.JobExecutionDtoCollection;
 import org.folio.rest.jaxrs.model.JobProfileInfo;
@@ -26,8 +29,11 @@ import org.springframework.stereotype.Repository;
 
 import javax.ws.rs.NotFoundException;
 import java.time.ZoneOffset;
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -62,6 +68,7 @@ import static org.folio.dao.util.JobExecutionDBConstants.SUBORDINATION_TYPE_FIEL
 import static org.folio.dao.util.JobExecutionDBConstants.TOTAL_COUNT_FIELD;
 import static org.folio.dao.util.JobExecutionDBConstants.TOTAL_FIELD;
 import static org.folio.dao.util.JobExecutionDBConstants.UI_STATUS_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.UPDATE_BY_IDS_SQL;
 import static org.folio.dao.util.JobExecutionDBConstants.UPDATE_SQL;
 import static org.folio.dao.util.JobExecutionDBConstants.USER_ID_FIELD;
 import static org.folio.rest.persist.PostgresClient.convertToPsqlStandard;
@@ -81,6 +88,18 @@ public class JobExecutionDaoImpl implements JobExecutionDao {
   private static final String TABLE_NAME = "job_execution";
   private static final String PROGRESS_TABLE_NAME = "job_execution_progress";
   public static final String GET_JOB_EXECUTION_HR_ID = "SELECT nextval('%s.job_execution_hr_id_sequence')";
+
+  //Below constants are used for building db query related to job execution deletions
+  public static final String ID = "id";
+  public static final String IS_DELETED = "is_deleted";
+  public static final String RETURNING_FIELD_NAMES = "returningFieldNames";
+  public static final String SET_CONDITIONAL_FIELD_VALUES = "setConditionalFieldValues";
+  public static final String SET_CONDITIONAL_FIELD_NAME = "setConditionalFieldName";
+  public static final String SET_FIELD_VALUE = "setFieldValue";
+  public static final String SET_FIELD_NAME = "setFieldName";
+  public static final String TENANT_NAME = "tenantName";
+  public static final String TRUE = "true";
+  public static final String DB_TABLE_NAME_FIELD = "tableName";
 
   @Autowired
   private PostgresClientFactory pgClientFactory;
@@ -203,6 +222,40 @@ public class JobExecutionDaoImpl implements JobExecutionDao {
         promise.complete(jobExecutionPromise.future().result());
       });
     return promise.future();
+  }
+
+  @Override
+  public Future<DeleteJobExecutionsResp> softDeleteJobExecutionsByIds(List<String> ids, String tenantId) {
+    Promise<RowSet<Row>> promise = Promise.promise();
+    try {
+      Map<String, String> data = new HashMap<>();
+      data.put(TENANT_NAME, convertToPsqlStandard(tenantId));
+      data.put(DB_TABLE_NAME_FIELD,TABLE_NAME);
+      data.put(SET_FIELD_NAME, IS_DELETED);
+      data.put(SET_FIELD_VALUE, TRUE);
+      data.put(SET_CONDITIONAL_FIELD_NAME, ID);
+      data.put(SET_CONDITIONAL_FIELD_VALUES, ids.stream().collect(Collectors.joining("','")));
+      data.put(RETURNING_FIELD_NAMES, ID + "," + IS_DELETED);
+      String query = StrSubstitutor.replace(UPDATE_BY_IDS_SQL, data);
+      pgClientFactory.createInstance(tenantId).execute(query, promise);
+    } catch (Exception e) {
+      LOGGER.error("Error deleting jobExecution by ids {}, ", ids, e);
+      promise.fail(e);
+    }
+    return promise.future().map(this::mapRowSetToDeleteChangeManagerJobExeResp);
+  }
+
+  private DeleteJobExecutionsResp mapRowSetToDeleteChangeManagerJobExeResp(RowSet<Row> rowSet){
+    DeleteJobExecutionsResp deleteJobExecutionsResp = new DeleteJobExecutionsResp();
+    List<JobExecutionDetail> jobExecutionDetails = new ArrayList<>();
+    rowSet.forEach(row -> {
+      JobExecutionDetail executionLogDetail = new JobExecutionDetail();
+      executionLogDetail.setJobExecutionId(row.getUUID(ID).toString());
+      executionLogDetail.setIsDeleted(row.getBoolean(IS_DELETED));
+      jobExecutionDetails.add(executionLogDetail);
+    });
+    deleteJobExecutionsResp.setJobExecutionDetails(jobExecutionDetails);
+    return deleteJobExecutionsResp;
   }
 
   private Tuple mapToTuple(JobExecution jobExecution) {