+++ mod-source-record-manager-server/src/main/java/org/folio/services/JournalRecordServiceImpl.java
@@ -3,11 +3,14 @@ package org.folio.services;
 import io.vertx.core.Future;
 import org.folio.dao.JournalRecordDao;
 import org.folio.rest.jaxrs.model.JobExecutionLogDto;
+import org.folio.rest.jaxrs.model.JobExecutionSummaryDto;
 import org.folio.rest.jaxrs.model.JobLogEntryDtoCollection;
 import org.folio.rest.jaxrs.model.JournalRecordCollection;
 import org.folio.rest.jaxrs.model.RecordProcessingLogDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 /**
  * Implementation of the JournalRecordService.
@@ -46,5 +49,10 @@ public class JournalRecordServiceImpl implements JournalRecordService {
   @Override
   public Future<RecordProcessingLogDto> getRecordProcessingLogDto(String jobExecutionId, String recordId, String tenantId) {
     return journalRecordDao.getRecordProcessingLogDto(jobExecutionId, recordId, tenantId);
+  }
+
+  @Override
+  public Future<Optional<JobExecutionSummaryDto>> getJobExecutionSummaryDto(String jobExecutionId, String tenantId) {
+    return journalRecordDao.getJobExecutionSummaryDto(jobExecutionId, tenantId);
   }
 }